SCRD-6919 Address layout problems with smaller screens
The styling attempted to visually merge the top title bar with the areas
below it (the left nav bar and the main content window), and to force
the left navbar to be drawn when it otherwise would not be. This caused
a number problems when viewing on a modest sized screen (1024 width) or
when resizing.

Removed some of the custom nav bar forced styling and let the
underlying styles of horizon and bootstrap correctly hide and show
the navbar when appropriate.  Also collapsed some of the repetitive
css into more friendly scss nested styles.